Classes moved into their own files.
Parameter names updated to describe more accurately.
Extension methods moved into their respective classes.
