+++ Modules/Intent.Modules.Common.CSharp/Builder/CSharpInterface.cs
@@ -132,9 +132,9 @@ public class CSharpInterface : CSharpDeclaration<CSharpInterface>, ICSharpRefere
         });
     }
     
-    public CSharpInterface AddMethod(CSharpType type, string name, Action<CSharpInterfaceMethod> configure = null)
+    public CSharpInterface AddMethod(CSharpType returnType, string name, Action<CSharpInterfaceMethod> configure = null)
     {
-        return InsertMethod(Methods.Count, type, name, configure);
+        return InsertMethod(Methods.Count, returnType, name, configure);
     }
 
     public CSharpInterface AddMethod(string returnType, string name, Action<CSharpInterfaceMethod> configure = null)
@@ -184,9 +184,9 @@ public class CSharpInterface : CSharpDeclaration<CSharpInterface>, ICSharpRefere
         return this;
     }
     
-    public CSharpInterface InsertMethod(int index, CSharpType type, string name, Action<CSharpInterfaceMethod> configure = null)
+    public CSharpInterface InsertMethod(int index, CSharpType returnType, string name, Action<CSharpInterfaceMethod> configure = null)
     {
-        var method = new CSharpInterfaceMethod(type, name, this)
+        var method = new CSharpInterfaceMethod(returnType, name, this)
         {
             BeforeSeparator = _methodsSeparator,
             AfterSeparator = _methodsSeparator
